+++ empla/integrations/hubspot/__init__.py
@@ -1,0 +1,1 @@
+"""empla.integrations.hubspot - HubSpot CRM connector."""